unify exception hierarchy
