+++ html/api/lib/errors.py
@@ -1,0 +1,16 @@
+class CurangelError(RuntimeError):
+    def __str__(self):
+        return self.fmt().capitalize()
+
+    def __repr__(self):
+        name = self.__class__.__name__
+        desc = self.fmt()
+        return "<{}: {}>".format(name, desc)
+        return self.__class__.__name__
+
+    def fmt(self, user=None):
+        if user is None:
+            user = "user"
+        else:
+            user = "user '{}'".format(user)
+        return self._fmt().strip().format(user=user, **vars(self))
